+++ Images/Converters/RAWImageConverter.cs
@@ -13,7 +13,6 @@ using System.Diagnostics.CodeAnalysis;
 using System.Diagnostics.Contracts;
 using System.Globalization;
 using System.Threading.Tasks;
-using OutputBuilderClient;
 using SixLabors.ImageSharp;
 using StorageHelpers;
