+++ Images/Images.csproj
@@ -1,29 +1,31 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
-    <PropertyGroup>
-        <TargetFramework>netstandard2.0</TargetFramework>
-        <LangVersion>latest</LangVersion>
-    </PropertyGroup>
-    <PropertyGroup>
-        <Company>Mark Ridgwell</Company>
-        <Copyright>Mark Ridgwell</Copyright>
-        <Product>Photo Gallery</Product>
-        <NeutralLanguage>en-GB</NeutralLanguage>
-    </PropertyGroup>
-    <PropertyGroup>
-        <AssemblyTitle>
-        </AssemblyTitle>
-        <Description>
-        </Description>
-    </PropertyGroup>
-    <ItemGroup>
-        <Folder Include="Properties"/>
-    </ItemGroup>
-    <ItemGroup>
-        <PackageReference Include="SixLabors.ImageSharp" Version="1.0.0-dev000605"/>
-        <PackageReference Include="SixLabors.ImageSharp.Drawing" Version="1.0.0-dev000605"/>
-    </ItemGroup>
-    <ItemGroup>
-        <ProjectReference Include="..\FileNaming\FileNaming.csproj"/>
-        <ProjectReference Include="..\ObjectModel\ObjectModel.csproj"/>
-    </ItemGroup>
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+    <LangVersion>latest</LangVersion>
+  </PropertyGroup>
+  <PropertyGroup>
+    <Company>Mark Ridgwell</Company>
+    <Copyright>Mark Ridgwell</Copyright>
+    <Product>Photo Gallery</Product>
+    <NeutralLanguage>en-GB</NeutralLanguage>
+  </PropertyGroup>
+  <PropertyGroup>
+    <AssemblyTitle>
+    </AssemblyTitle>
+    <Description>
+    </Description>
+  </PropertyGroup>
+  <ItemGroup>
+    <Folder Include="Properties"/>
+  </ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="ExifLib" Version="1.7.0"/>
+    <PackageReference Include="SixLabors.ImageSharp" Version="1.0.0-dev000605"/>
+    <PackageReference Include="SixLabors.ImageSharp.Drawing" Version="1.0.0-dev000605"/>
+    <PackageReference Include="Taglib-Sharp-NetStandard" Version="1.0.0"/>
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\FileNaming\FileNaming.csproj"/>
+    <ProjectReference Include="..\ObjectModel\ObjectModel.csproj"/>
+  </ItemGroup>
 </Project>
